requirements.txt
streamlit
openai
python-dotenv
llama-index
numpy
scikit-learn
chromadb
protobuf==3.20.1
pysqlite3-binary